+++ CHANGELOG.md
@@ -42,7 +42,8 @@ Changelog
 * Made *accounts/hasFeatures* dangerous
 * Added restrictions [SMS_DELIVER](https://developer.android.com/reference/android/provider/Telephony.Sms.Intents.html#SMS_DELIVER_ACTION) and [WAP_PUSH_DELIVER](https://developer.android.com/reference/android/provider/Telephony.Sms.Intents.html#WAP_PUSH_DELIVER_ACTION)
 * Corrected some restriction permissions
-* Moved *getAllByName*, *getByAddress* and *getByName* ([InetAddress](http://developer.android.com/reference/java/net/InetAddress.html)) from *internet* to *network*
+* Moved restrictions *getAllByName*, *getByAddress* and *getByName* ([InetAddress](http://developer.android.com/reference/java/net/InetAddress.html)) from *internet* to *network*
+* Fixed restrictions *getByAddress* and *getByName*; to prevent problems these dangerous restrictions need to be set again
 * Updated in application documentation
 * Updated Dutch translation
 * Updated German translation